pyproject.toml
[project]
name = "pge-etl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.42.49",
    "polars>=1.38.1",
    "psycopg2>=2.9.11",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
]
